+++ dlls/ai_playerally.cpp
@@ -133,7 +133,7 @@ static CUtlMap<AIConcept_t, ConceptInfo_t *> g_ConceptInfoMap;
 static void InitConcepts( void )
 {
 	g_ConceptInfoMap.SetLessFunc( CaselessStringLessThan );
-	for ( int i = 0; i < ARRAYSIZE(g_ConceptInfos); i++ )
+	for ( size_t i = 0; i < ARRAYSIZE(g_ConceptInfos); i++ )
 	{
 		g_ConceptInfoMap.Insert( g_ConceptInfos[i].concept, &g_ConceptInfos[i] );
 	}
@@ -158,7 +158,7 @@ void CAI_AllySpeechManager::Spawn()
 		InitConcepts();
 	}
 
-	for ( int i = 0; i < ARRAYSIZE(g_ConceptInfos); i++ )
+	for ( size_t i = 0; i < ARRAYSIZE(g_ConceptInfos); i++ )
 		m_ConceptTimers.Insert( AllocPooledString( g_ConceptInfos[i].concept ), CSimpleSimTimer() );
 }
 
@@ -1512,7 +1512,7 @@ bool CAI_PlayerAlly::ShouldSpeakRandom( AIConcept_t concept, int iChance )
 			if ( flModifier < 0.001 )
 				return false;
 				
-			iChance = floor( (float)iChance / flModifier );
+			iChance = static_cast<int>(floor( (float)iChance / flModifier ));
 		}
 	}
