+++ dlls/bmodels.cpp
@@ -701,7 +701,7 @@ void CFuncRotating::RampPitchVol( void )
 
 	float fpitch = FANPITCHMIN + (FANPITCHMAX - FANPITCHMIN) * fpct;	
 	
-	int pitch = clamp(fpitch, 0, 255);
+	int pitch = static_cast<int>(clamp(fpitch, 0, 255));
 	if (pitch == PITCH_NORM)
 	{
 		pitch = PITCH_NORM - 1;
