+++ dlls/buttons.cpp
@@ -141,19 +141,19 @@ bool CBaseButton::KeyValue( const char *szKeyName, const char *szValue )
 {
 	if (FStrEq(szKeyName, "locked_sound"))
 	{
-		m_bLockedSound = atof(szValue);
+		m_bLockedSound = static_cast<byte>(atof(szValue));
 	}
 	else if (FStrEq(szKeyName, "locked_sentence"))
 	{
-		m_bLockedSentence = atof(szValue);
+		m_bLockedSentence = static_cast<byte>(atof(szValue));
 	}
 	else if (FStrEq(szKeyName, "unlocked_sound"))
 	{
-		m_bUnlockedSound = atof(szValue);
+		m_bUnlockedSound = static_cast<byte>(atof(szValue));
 	}
 	else if (FStrEq(szKeyName, "unlocked_sentence"))
 	{
-		m_bUnlockedSentence = atof(szValue);
+		m_bUnlockedSentence = static_cast<byte>(atof(szValue));
 	}
 	else
 	{
