+++ dlls/fire.cpp
@@ -940,7 +940,7 @@ void CFire::Update( float simTime )
 	if ( m_flDamageTime <= gpGlobals->curtime )
 	{
 		m_flDamageTime = gpGlobals->curtime + fire_dmginterval.GetFloat();
-		outputDamage = (fire_dmgbase.GetFloat() + outputHeat * fire_dmgscale.GetFloat() * m_flDamageScale) * fire_dmginterval.GetFloat();
+		outputDamage = static_cast<int>((fire_dmgbase.GetFloat() + outputHeat * fire_dmgscale.GetFloat() * m_flDamageScale) * fire_dmginterval.GetFloat());
 		if ( outputDamage )
 		{
 			damage = true;
@@ -957,7 +957,7 @@ void CFire::Update( float simTime )
 		}
 		else if ( FClassnameIs( pOther, "env_fire" ) )
 		{
-			if ( fireCount < ARRAYSIZE(pFires) )
+			if ( fireCount < static_cast<int>(ARRAYSIZE(pFires)) )
 			{
 				pFires[fireCount] = (CFire *)pOther;
 				fireCount++;
