+++ dlls/hl2_dll/basehlcombatweapon.cpp
@@ -100,7 +100,7 @@ void CHLMachineGun::PrimaryAttack( void )
 	//Factor in the view kick
 	AddViewKick();
 
-	CSoundEnt::InsertSound( SOUND_COMBAT, GetAbsOrigin(), SOUNDENT_VOLUME_MACHINEGUN, 0.2, pPlayer );
+	CSoundEnt::InsertSound( SOUND_COMBAT, GetAbsOrigin(), static_cast<int>(SOUNDENT_VOLUME_MACHINEGUN), 0.2, pPlayer );
 	
 	if (!m_iClip1 && pPlayer->GetAmmoCount(m_iPrimaryAmmoType) <= 0)
 	{
